refactor(KERN): remove legacy design split date input
