+++ stories/form/inputs/SplitDateInput.stories.tsx
@@ -1,7 +1,7 @@
 import { reactRouterFormContext } from "~/../.storybook/reactRouterFormContext";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import z from "zod";
-import KernSplitDateInput from "~/components/kern/formElements/input/KernSplitDateInput";
+import KernSplitDateInput from "~/components/formElements/KernSplitDateInput";
 import { Grid } from "~/components/layout/grid/Grid";
 import { GridItem } from "~/components/layout/grid/GridItem";
 import { GridSection } from "~/components/layout/grid/GridSection";
